refactor: rename saveVisionBoardData to saveDreamBoardData, remove AI_GENERATED content type, and update related components for improved clarity and consistency
