+++ src/features/dream-board/components/Board/DreamBoardContainer/index.ts
@@ -1,4 +1,3 @@
 export { DreamBoard } from '../Board';
-export * from '../types';
 export * from '../../../utils/boardUtils';
 export { DreamBoardContainer } from './DreamBoardContainer';
